backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "gerador de certificado de morte, beetlejuice"
authors = ["shiizuko <juliashizuko66@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
pillow = "^10.4.0"
supabase = "^2.7.4"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
